+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/passport/decrypted/abstracts/SecureValueWithData.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.types.passport.decrypted.abstracts
+
+import dev.inmo.tgbotapi.types.passport.credentials.DataCredentials
+
+interface SecureValueWithData : SecureValue {
+    val data: DataCredentials?
+}