+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/passport/encrypted/abstracts/FilesCollection.kt
@@ -1,0 +1,10 @@
+package dev.inmo.tgbotapi.types.passport.encrypted.abstracts
+
+import dev.inmo.tgbotapi.types.passport.encrypted.EncryptedElementSerializer
+import dev.inmo.tgbotapi.types.passport.encrypted.PassportFile
+import kotlinx.serialization.Serializable
+
+@Serializable(EncryptedElementSerializer::class)
+interface FilesCollection : EncryptedPassportElement {
+    val files: List<PassportFile>
+}