+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/passport/encrypted/abstracts/WithEmail.kt
@@ -1,0 +1,9 @@
+package dev.inmo.tgbotapi.types.passport.encrypted.abstracts
+
+import dev.inmo.tgbotapi.types.passport.encrypted.EncryptedElementSerializer
+import kotlinx.serialization.Serializable
+
+@Serializable(EncryptedElementSerializer::class)
+interface WithEmail : EncryptedPassportElement {
+    val email: String
+}