+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/passport/encrypted_data/abstracts/WithPhoneNumber.kt
@@ -1,9 +1,0 @@
-package dev.inmo.tgbotapi.types.passport.encrypted_data.abstracts
-
-import dev.inmo.tgbotapi.types.passport.encrypted_data.EncryptedElementSerializer
-import kotlinx.serialization.Serializable
-
-@Serializable(EncryptedElementSerializer::class)
-interface WithPhoneNumber : EncryptedPassportElement {
-    val phoneNumber: String
-}