+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 * `Common`:
     * `Version`:
-        * `MicroUtils`: `0.4.16` -> `0.4.21`
+        * `MicroUtils`: `0.4.16` -> `0.4.23`
+        * `Klock`: `0.2.3` -> `0.2.4`
+        * `Ktor`: `1.5.0` -> `1.5.1`
 * `Core`:
     * **BREAKING CHANGE** Now `MediaGroupMessage` have a generic type related to `MediaGroupContent`
         * Methods and types related to `MediaGroupMessage` have been modified according to their meanings
@@ -19,6 +21,8 @@
     * `onVisualMediaGroup` now is just an alternative to `onVisualGallery`
     * `command` and `onCommand` expectations has been added for commands `String` variant
     * New extensions `BehaviourContext#oneOf`, `BehaviourContext#parallel` and `Deferred<T>#withAction`
+* `API`:
+    * **PASSPORT** New extensions `TelegramBot#setPassportDataErrors`
 
 ## 0.31.0